fixed most views except filter table view
