+++ public/components/graph/Graph.js
@@ -21,12 +21,13 @@ import { ReactiveFormCreate } from "./graphComponents/ReactiveFormCreate.js";
 import { FilterBox} from "../filter/FilterBox.js"
 import { checkFilter } from "../filter/filterFunctions.js"
 
-async function Graph(view, parentNode) {
+async function Graph(view, useFilterPara = false, parentNode) {
   State.view = view;
   Actions.GETDEF();
 
   let nodes,
     rels = [];
+  let divContainer;
   let graphJsonData = await JSON.parse(sessionStorage.getItem(view));
 
   nodes = graphJsonData[0].nodes;
@@ -342,7 +343,6 @@ async function Graph(view, parentNode) {
   };
 
   let link = g.append("g").attr("class", "forLinks").select(".forLinks");
-  // .selectAll("path")
 
   let linkLabel = g
     .selectAll(".linkLabel")
@@ -443,13 +443,15 @@ async function Graph(view, parentNode) {
 
   async function render(view) {
     await updateData(view);
-    d3.select("#filter-container-id").remove()
-    divContainer = d3.select(await FilterBox())
-    parentNode.appendChild(divContainer.node());
-    divContainer.selectAll(".form-check-input").on("click", async function(e) {
-      await checkFilter(e)
-      await render("filter");
-    })
+    if(useFilterPara){
+      d3.select("#filter-container-id").remove()
+      divContainer = d3.select(await FilterBox())
+      parentNode.appendChild(divContainer.node());
+      divContainer.selectAll(".form-check-input").on("click", async function(e) {
+        await checkFilter(e)
+        await render("filter");
+      })
+    }
     simulation.stop();
 
     link = svg
@@ -599,46 +601,14 @@ async function Graph(view, parentNode) {
     simulation.alpha(1).restart();
     return svg.node()
   }
-  let divContainer = d3.select(await FilterBox())
-  await render(view);
-/*
-  let divContainer = d3.select(await FilterBox())
 
-  let divContainerCheck = divContainer.selectAll(".form-check-input").on("click", async function(e) {
-     //d3.select(".accordion-body").append(d3.select(await triggerTreeGetHtml()).node())
-    await checkFilter(e)
-    //console.log("CLICK")
-    //d3.select("#filter-container-id").remove()
-    await render("filter");
-    // d3.select(".filter-container").remove();
-    // d3.select(await FilterBox())
-    // console.log(e)
-
-  })
-*/
-
-  // console.log(divContainerCheck)
-  // let divContainer =   d3.select(await FilterBox()).attr("id", "TEST").selectAll(".form-check-input").on("change", (d) => console.log(d, "click!"))
-
-  // console.log(  d3.select(await FilterBox()).selectAll(".form-check-input").on("click", (d) => console.log(d, "click!")),  "filter")
-
-  // console.log(d3.select("#accordionPanelsStayOpenExample"))
-  // let divContainer = d3.select(await FilterBox()).on("click", () => console.log("click!"))
-
-
-  // let divContainer = d3.select("#app")
-  // .append("div")
-  // .attr("class","filterContainer")
-  // .select(".filterbutton")
-  // .on("click", () => console.log("hello"))
-
-  
-  // .append("button")
-  // .attr("id", "testbutton")
-  
- 
-  return [svg.node(), divContainer.node()]
-
-  // return svg.node();
+  if(useFilterPara){
+    divContainer = d3.select(await FilterBox())
+    await render(view);
+    return [svg.node(), divContainer.node()]
+  } else{
+    await render(view);
+    return [svg.node()]
+  }
 }
 export default Graph;
