+++ public/views/Filter.js
@@ -4,7 +4,7 @@ import {FilterBox} from "../components/filter/FilterBox.js";
 import Actions from "../store/Actions.js";
 import { renderDataAsGraph, setupToolBar } from "../components/table/dataRendererHelper.js";
 import {State} from "../store/State.js";
-import {createHtmlElementWithData, appendChildsToSelector} from "../components/table/dataRendererHelper.js";
+import {createHtmlElementWithData} from "../components/DomElementHelper.js";
 
 export default class Filter {
     constructor() {
@@ -25,12 +25,12 @@ export default class Filter {
 
         sessionStorage.setItem("filter", JSON.stringify([{nodes: tree.selectedNodesData , rels: tree.visibleRelations }]));
         let rowDiv = createHtmlElementWithData("div", {"class": "row"})
-        let firstColumnDiv = createHtmlElementWithData("div", {"class": "col-md", "id": "filterbox-grid-container-id"})
+        let firstColumnDiv = createHtmlElementWithData("div", {"class": "col-sm", "id": "filterbox-grid-container-id"})
         rowDiv.appendChild(firstColumnDiv)
-        let secondColumnDiv = createHtmlElementWithData("div", {"class": "col-md", "id": "data-display-grid-container-id"})
+        let secondColumnDiv = createHtmlElementWithData("div", {"class": "col-lg", "id": "data-display-grid-container-id"})
         rowDiv.appendChild(secondColumnDiv)
 
-        const mainAppNodes = await this.returnRenderFunc("filter", firstColumnDiv)
+        const mainAppNodes = await this.returnRenderFunc("filter", true, firstColumnDiv)
         firstColumnDiv.appendChild(mainAppNodes[1])
         secondColumnDiv.appendChild(mainAppNodes[0])
         return [rowDiv]
